ext/flac/: Cast some size_t arguments to guint to avoid compiler warnings on 64-bit systems.
Original commit message from CVS:
* ext/flac/gstflacdec.c (gst_flac_dec_read_stream):
* ext/flac/gstflacenc.c (gst_flac_enc_write_callback):
Cast some size_t arguments to guint to avoid compiler
warnings on 64-bit systems.

+++ ext/flac/gstflacdec.c
@@ -982,7 +982,7 @@ gst_flac_dec_read_stream (const FLAC__StreamDecoder * decoder,
   }
 
   GST_LOG_OBJECT (dec, "feeding %u bytes to decoder (available=%u, bytes=%u)",
-      len, gst_adapter_available (dec->adapter), *bytes);
+      len, gst_adapter_available (dec->adapter), (guint) * bytes);
   memcpy (buffer, gst_adapter_peek (dec->adapter, len), len);
   *bytes = len;
